+++ src/main/java/faang/school/achievement/config/WriterAchievementConfig.java
@@ -2,6 +2,7 @@ package faang.school.achievement.config;
 
 import faang.school.achievement.handler.WriterAchievementHandler;
 import faang.school.achievement.service.AchievementService;
+import faang.school.achievement.service.UserAchievementService;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.beans.factory.annotation.Value;
@@ -13,12 +14,13 @@ import org.springframework.context.annotation.Configuration;
 public class WriterAchievementConfig {
 
     private final AchievementService achievementService;
+    private final UserAchievementService userAchievementService;
 
     @Value("${achievement.writer.points}")
     private int writerPoints;
 
     @Bean
     public WriterAchievementHandler writerAchievementHandler() {
-        return new WriterAchievementHandler(achievementService, writerPoints);
+        return new WriterAchievementHandler(achievementService, userAchievementService, writerPoints);
     }
 }
