+++ src/main/java/faang/school/achievement/handler/WriterAchievementHandler.java
@@ -3,6 +3,7 @@ package faang.school.achievement.handler;
 import faang.school.achievement.dto.PostEventDto;
 import faang.school.achievement.model.AchievementTitle;
 import faang.school.achievement.service.AchievementService;
+import faang.school.achievement.service.UserAchievementService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 
@@ -10,12 +11,14 @@ import org.springframework.scheduling.annotation.Async;
 @RequiredArgsConstructor
 public class WriterAchievementHandler {
     private final AchievementService achievementService;
+    private final UserAchievementService userAchievementService;
+
     private final long requiredPoints;
 
     @Async("threadPool")
     public void computeAchievement(PostEventDto postEventDto) {
         Long userId = postEventDto.getAuthorId();
-        if (!achievementService.hasAchievement(userId, AchievementTitle.WRITER)) {
+        if (!userAchievementService.hasAchievement(userId, AchievementTitle.WRITER)) {
             achievementService.updateProgress(userId, AchievementTitle.WRITER, requiredPoints);
         }
     }
