+++ src/test/java/faang/school/achievement/service/AchievementServiceTest.java
@@ -6,7 +6,6 @@ import faang.school.achievement.model.AchievementTitle;
 import faang.school.achievement.model.UserAchievement;
 import faang.school.achievement.repository.AchievementProgressRepository;
 import faang.school.achievement.repository.AchievementRepository;
-import faang.school.achievement.repository.UserAchievementRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -31,10 +30,10 @@ class AchievementServiceTest {
     private AchievementRepository achievementRepository;
 
     @Mock
-    private UserAchievementRepository userAchievementRepository;
+    private AchievementProgressRepository achievementProgressRepository;
 
     @Mock
-    private AchievementProgressRepository achievementProgressRepository;
+    private UserAchievementService userAchievementService;
 
     @InjectMocks
     private AchievementService achievementService;
@@ -60,19 +59,20 @@ class AchievementServiceTest {
 
     @Test
     void testHasAchievement_True() {
-        when(userAchievementRepository.hasAchievement(userId, title)).thenReturn(true);
 
-        boolean result = achievementService.hasAchievement(userId, title);
+        when(userAchievementService.hasAchievement(userId, title)).thenReturn(true);
+
+        boolean result = userAchievementService.hasAchievement(userId, title);
 
         assertTrue(result);
-        verify(userAchievementRepository, times(1)).hasAchievement(userId, title);
+        verify(userAchievementService, times(1)).hasAchievement(userId, title);
     }
 
     @Test
     void testHasAchievement_False() {
-        achievementService.hasAchievement(userId, title);
+        userAchievementService.hasAchievement(userId, title);
 
-        verify(userAchievementRepository, times(1)).hasAchievement(userId, title);
+        verify(userAchievementService, times(1)).hasAchievement(userId, title);
     }
 
     @Test
@@ -92,11 +92,11 @@ class AchievementServiceTest {
     void testUpdateProgress_Achieved() {
         when(achievementProgressRepository.getProgress(userId, title)).thenReturn(Optional.of(progress));
         when(achievementRepository.findByTitle(title)).thenReturn(Optional.of(achievement));
-
+        when(userAchievementService.createAndSaveNewUserAchievement(progress)).thenReturn(mock(UserAchievement.class));
         achievementService.updateProgress(userId, title, requiredPoints);
 
         assertEquals(10, progress.getCurrentPoints());
-        verify(userAchievementRepository, times(1)).save(any(UserAchievement.class));
+        verify(userAchievementService, times(1)).createAndSaveNewUserAchievement(progress);
     }
 
     @Test
